+++ database/migrations/2025_01_08_073446_create_employe_salaries_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->ulid('id')->primary();
             $table->ulid('employe_id');
             $table->ulid('salary_id');
-            $table->integer('basic_salary')->default(0);
-            $table->integer('meal_allowances')->default(0);
-            $table->integer('transport_allowances')->default(0);
-            $table->integer('position_allowances')->default(0);
+            $table->numeric('basic_salary')->default(0);
+            $table->numeric('meal_allowances')->default(0);
+            $table->numeric('transport_allowances')->default(0);
+            $table->numeric('position_allowances')->default(0);
             $table->boolean('is_using_master_salary')->default(true);
             $table->date('periode');
             $table->boolean('is_active')->default(true);
